+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/DateTime.kt
@@ -1,0 +1,12 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions
+
+import org.joda.time.DateTime
+import org.joda.time.DateTimeZone
+
+fun DateTime.withoutTimeZoneOffset(): DateTime {
+    return withZone(
+        DateTimeZone.UTC
+    ).plus(
+        zone.toTimeZone().rawOffset.toLong()
+    )
+}